Basic login to twitch irc and read messages
