+++ .gitignore
@@ -1,2 +1,3 @@
 __pycache__/*
 actions/__pycache__/*
+.env
