+++ default.env
@@ -1,0 +1,5 @@
+# Write the config values to the right of the equals sign, with no space inbetween.
+# Your Twitch username
+TWITCHUSERNAME=
+# Your Twitch Chat OAuth token (from https://twitchapps.com/tmi/)
+TWITCHTOKEN=
